requirements.txt
aiogram==3.0.0b6
aiohttp
pymongo
datetime
asyncio
redis
python-dotenv
cachetools
arq
pathLib
Babel
PyBabel
apscheduler
apscheduler_di
uvicorn
fastapi
hashlib
dacite
python-dateutil